FIX if..then and add better install options

+++ setup_scripts/1-prep.sh
@@ -1,16 +1,34 @@
 #!/bin/bash
 
-if ! $(zenity --question --title "Are you Crazy?" --text="Danger!  Do you really want to continue?  Be aware this script destructively loosens permissions for the current user, changes system settings, uninstalls *many* applications, and installs *many* applications not supported by Cannonical (the makers of Ubuntu).
+## Prompt to give user a chance to abort to avoid borking their system
+INSTALLTYPE=$(zenity  --list \
+  --radiolist  \
+  --timeout=30 \
+  --hide-column=2 \
+  --column "" TRUE abort "Abort Installation" FALSE virtual "Install Virtual Kernel" FALSE standard "Install Standard Kernel (physical hardware)" \
+  --column value \
+  --column Description \
+  --width=400 \
+  --title="Danger!" \
+  --text="\
+Danger!  Do you really want to continue?  Be aware this script destructively
+loosens permissions for the current user, changes system settings, uninstalls
+*many* applications, and installs *many* applications not supported by Cannonical
+(the makers of Ubuntu).
 
-Are you really, really, ... and I mean *really* sure you want to do this?")
+Are you really, really, ... and I mean *really* sure you want to do this?" \
+)
+
+if [ $? != 0 ] || [ "$INSTALLTYPE" == "abort" ]
 then
   echo "aborting";
-  exit -1;
-else
-zenity --info --text="This script may take hours to run (depending on your hardware and internet bandwidth), plus multiple automated reboots.
+  exit 1;
+fi
 
-Towards the end, the process requires some manual steps, guided by popups like this. \nThis script shouldn't be run more than once.  NOTE:  This script should be run by user: quickstart -- and should be setup to allow automatic login upon reboot";
+# ok, lets do it
+zenity --question --text="This script may take hours to run (depending on your hardware and internet bandwidth), plus multiple automated reboots (which requires AUTOMATIC USER LOGIN for an unattended setup).
 
+Towards the end, the process requires some manual steps, guided by popups like this. \nThis script shouldn't be run more than once.";
 
 ## The last password you'll ever need.
 # add current user to sudoers file - careful, this line could brick the box.
@@ -21,19 +39,19 @@ sudo chmod 440 /etc/sudoers.d/quickstart
 # note: seems unsafe for anyone unaware.
 sudo adduser $USER root
 
-
 ## Disk size Accounting
 # Starting size:
 df -h -T > ~/quickstart/setup_scripts/logs/quickstart-size-start.txt
-
 
 ## Some configuration
 # turn off screen saver
 gconftool-2 -s /apps/gnome-screensaver/idle_activation_enabled --type=bool false
 
-# Install virtual kernel.  Better performance.
-sudo apt-get -yq update
-sudo apt-get -yq purge linux-generic linux-headers-generic linux-image-generic linux-generic-pae  linux-image-generic-pae linux-headers-generic-pae linux-headers-3.2.0-23 linux-headers-3.2.0-23-generic-pae linux-image-3.2.0-23-generic-pae
-sudo apt-get -yq install linux-virtual linux-headers-virtual linux-image-virtual
-
+if [ "$INSTALLTYPE" == "virtual" ]
+then
+  # Remove the standard Kernel and install virtual kernel.
+  # Should be less overhead (aka better performance) in Virtual environment.
+  sudo apt-get -yq update
+  sudo apt-get -yq purge linux-generic linux-headers-generic linux-image-generic linux-generic-pae  linux-image-generic-pae linux-headers-generic-pae linux-headers-3.2.0-23 linux-headers-3.2.0-23-generic-pae linux-image-3.2.0-23-generic-pae
+  sudo apt-get -yq install linux-virtual linux-headers-virtual linux-image-virtual
 fi
